==> TrifosNineI2/5_TrifoIv2_Calc5Graphs.Rmd <==
---
title: "Emergence seedlings I2 graphs with Mean Soil Temp"
author: "CT"
date: "17.08.2017"
output: 
  word_document: default
  pdf_document: default
  html_document: default
---

This script does this:
-get dataframe produced in Calc4 (df_MeanTempNineTrifo.txt)
-graph y=Days to 9 trifoliate ; x= mean air Temp
-graph y= 1/days to 9trifoliates ; x= mean air Temp
-produces the linear equations leaf appearance  as a function of temperature 



```{r loadLibraries, warning=FALSE}
library(tidyr)
library(dplyr)
library(ggplot2)
library(agricolae)
library(knitr)
library(lubridate)
library(mosaic)
library(lattice)
library (gridExtra)
```

Read file and correct formats 
```{r loadFile}
# Get dataframe from previous script (written table )
setwd("C:\\GitHubRepos\\SubCloverPhD\\SubClover\\TrifosNineI2")
getwd()

#create file
#df_G <- read.table("df_DAS5Trifo_MeanAirT.txt",header=TRUE)

df_G <- read.table("df_MeanTempNineTrifo.txt",header=TRUE)




```

```{r}
df_G$DAS9T <- as.numeric(df_G$DAS9T)
df_G$SowingDate <- ymd(df_G$SowingDate )
df_G$Cultivar <- as.factor(df_G$Cultivar)
df_G$Plot <- as.factor(df_G$Plot)


#check data 
str(df_G)


# this is to "clean previous dfs : df_emerg4 <- NULL
```


```{r}

df_mean<-df_G%>%
  group_by(Cultivar,SowTreat)%>%
   summarise_all(funs(mean))

str(df_mean)

```



Graph DAS9T  and rate against meanTemp at trifoliate appearance   do soil Temp first 

```{r, fig.height=5, fig.width=10, echo = FALSE}



fontSize <- 12.0


  g1 <- df_mean %>%
  ggplot(aes(x= meanTempSoil_9T, y=DAS9T))  +
  geom_point(aes(shape= SowTreat), size=4)+
  scale_shape_manual(values = c(0,1,2,3,4,5,6,7,8))+
  #geom_smooth()+
  #stat_smooth(level = 0.9, size = 1, colour="white") + 
  theme_bw() +
  xlab(expression(Mean~Soil~temperature~~"("^"o"*"C)")) +  
  ylab("Days to nive trifoliates  (days)") +
    facet_grid(.~Cultivar, scales="free")+
  theme(text = element_text(size=11))+
  scale_x_continuous(limits = c(0,25))+
  scale_y_continuous(limits = c(0,120))

# 1/DAS9T
  g2 <- df_mean %>%
  ggplot(aes(x= meanTempSoil_9T, y=Rate, shape=SowTreat))  +
  geom_point(aes(shape= SowTreat), size=4)+
  scale_shape_manual(values = c(0,1,2,3,4,5,6,7,8))+
  theme_bw() +
  xlab(expression(Mean~Soil~temperature~~"("^"o"*"C)")) + 
  ylab(" Nine Trifoliates Rate (1/ Days) ") +
  facet_grid(.~Cultivar, scales="free")+
  theme(text = element_text(size=11))+
 scale_x_continuous(limits = c(0,25))+
 scale_y_continuous(limits = c(0,0.03))
  
  #grid.arrange(g1, g2)
# 
g <-grid.arrange(g1, g2)
  
  ggsave(file="plot_DAS_Rate_9TrifoIversenSoilTemp.png", width=12, height=6, dpi=400,g)
# 

```

Do same graph also with air temperatures

Graph DAS9T  and rate against Air meanTemp at trifoliate appearance (Air temp now)

```{r, fig.height=5, fig.width=10, echo = FALSE}


fontSize <- 12.0


  g3 <- df_mean %>%
  ggplot(aes(x= meanTempAir_9T, y=DAS9T))  +
  geom_point(aes(shape= SowTreat), size=4)+
  scale_shape_manual(values = c(0,1,2,3,4,5,6,7,8))+
  #geom_smooth()+
  #stat_smooth(level = 0.9, size = 1, colour="white") + 
  theme_bw() +
  xlab(expression(Mean~Air~temperature~~"("^"o"*"C)")) +  
  ylab("Days to nine trifoliates (days)") +
    facet_grid(.~Cultivar, scales="free")+
  theme(text = element_text(size=11))+
  scale_x_continuous(limits = c(0,25))+
  scale_y_continuous(limits = c(0,120))

# 1/DAS9T
  g4 <- df_mean %>%
  ggplot(aes(x= meanTempAir_9T, y=Rate, shape=SowTreat))  +
  geom_point(aes(shape= SowTreat), size=4)+
  scale_shape_manual(values = c(0,1,2,3,4,5,6,7,8))+
  theme_bw() +
  xlab(expression(Mean~Air~temperature~~"("^"o"*"C)")) + 
  ylab(" Nine Trifoliates Rate (1/ Days) ") +
  facet_grid(.~Cultivar, scales="free")+
  theme(text = element_text(size=11))+
 scale_x_continuous(limits = c(0,25))+
 scale_y_continuous(limits = c(0,0.03))
  
  #grid.arrange(g1, g2)
# 
g <-grid.arrange(g3, g4)
  
  ggsave(file="plot_DAS_Rate_9TrifoIversenAirTemp.png", width=12, height=6, dpi=400,g)
# 

```




Then get the equations from the whole temperature range  in all sowing dates  



use the means for each sowtreat 



calculate mean values for each cultivar and sow Treat

```{r}
df_mean<-df_G%>%
  group_by(Cultivar,SowTreat)%>%
   summarise_all(funs(mean,sd))
```

```{r, fig.height=5, fig.width=10, echo = FALSE}
#DAS50E
#
  g5 <- df_mean %>%
  #filter(Cultivar=="Monti"| Cultivar=="Antas")%>%
 # filter (SowTreat== "S3"| SowTreat== "S4"|SowTreat== "S5"|SowTreat== "S6")%>%
    #group_by(Cultivar,SowTreat)%>%
    #summarise_all(funs(mean))
  ggplot(aes(x=MeanAirTempTrifo_mean, y=DAS5T_mean, colour=SowTreat))  +
    geom_point(aes(shape= SowTreat), size=4)+
  scale_shape_manual(values = c(0,1,2,3,4,5,6,7,8))+
  
  #geom_errorbar(aes(ymin=DAS5T_mean-DAS5T_sd/2,
  #ymax=DAS5T_mean+DAS5T_sd/2),width=3,colour="grey")+
  #geom_errorbar(aes(ymin=2-DAS5T_sd/2,
  #ymax=2+DAS5T_sd/2),width=2,colour="grey")+
 # geom_point(alpha=0.8) +
  #geom_smooth()+
  #stat_smooth(level = 0.9, size = 1, colour="white") + 
  theme_bw() +
  xlab(expression(Mean~Air~temperaturee~~"("^"o"*"C)")) +  
  ylab("Days to 5 trifoliates  (days)") +
    facet_grid(.~Cultivar, scales="free")+
  theme(text = element_text(size=11))+
  scale_x_continuous(limits = c(0,20))+
  scale_y_continuous(limits = c(0,90))

# 1/DAS50E
  g6 <- df_mean %>%
   #filter(Cultivar=="Monti"| Cultivar=="Antas")%>%
  #filter (SowTreat== "S3"| SowTreat== "S4"|SowTreat== "S5"|SowTreat== "S6")%>%
  ggplot(aes(x=MeanAirTempTrifo_mean, y=Rate_mean, colour=SowTreat))  +
    geom_point(aes(shape= SowTreat), size=4)+
  scale_shape_manual(values = c(0,1,2,3,4,5,6,7,8))+
  #geom_point(alpha=0.8) +
  #geom_smooth(colour="red")+
  #stat_smooth(level = 0.9, size = 1, colour="white") + 
  theme_bw() +
  xlab(expression(Mean~Air~temperature~~"("^"o"*"C)")) + 
  ylab(" Rate (1/ Days) to 5 trifoliates") +
  facet_grid(.~Cultivar, scales="free")+
  theme(text = element_text(size=11))+
 scale_x_continuous(limits = c(0,20))+
  scale_y_continuous(limits = c(0,0.04))
  grid.arrange(g3, g4)
```






```{r, fig.height=7,fig.width=12}
require(ggplot2)
require(plyr)
require(dplyr)

str(df_G)

lm_eqn = function(df){
  m = lm(y ~ x, df);
  eq <- substitute(italic(y) == a + b %.% italic(x)*","~~italic(r)^2~"="~r2, 
                   list(a = format(coef(m)[1], digits = 2), 
                        b = format(coef(m)[2], digits = 2), 
                        r2 = format(summary(m)$r.squared, digits = 3)))
  as.character(as.expression(eq));                 
}


df <- df_G %>%
  dplyr::select(SowTreat,Cultivar, Plot, meanTempAir_9T, Rate) %>%
  group_by(SowTreat,Cultivar) %>%
  summarise_each(funs(mean)) %>%
  mutate(y=Rate,x=meanTempAir_9T) %>%
  dplyr::select(-meanTempAir_9T, -Rate)

eq <- ddply(df,.(Cultivar),lm_eqn)

yx<-quantile(df$y,0.75)
xn<- quantile(df$x,0.25)

#get statistical summary here 
summary(lm(y ~ x, df))

p <- ggplot(data = df, aes(x = x, y = y)) +
  geom_smooth(method = "lm", se=FALSE, color="black", formula = y ~ x) +
  geom_point(aes(colour=SowTreat),size=4)+
  theme_bw()+
  scale_x_continuous(limits = c(0,25))+
  scale_y_continuous(limits = c(0,0.045))

p1 <- p + geom_text(data=eq,aes(x = xn, y = yx,label=V1), parse = TRUE, inherit.aes=FALSE) + 
  facet_grid(Cultivar~.,scales="free")
  
p1
```

Do it individually for each cultivar
```{r}

require(ggplot2)
require(plyr)
require(dplyr)

lm_eqn = function(df){
  m = lm(y ~ x, df);
  eq <- substitute(italic(y) == a + b %.% italic(x)*","~~italic(r)^2~"="~r2, 
                   list(a = format(coef(m)[1], digits = 2), 
                        b = format(coef(m)[2], digits = 2), 
                        r2 = format(summary(m)$r.squared, digits = 3)))
  as.character(as.expression(eq));                 
}


df <- df_G %>%
  dplyr::select(SowTreat,Cultivar, Plot, meanTempAir_9T, Rate) %>%
  group_by(SowTreat,Cultivar) %>%
  filter(Cultivar=='Denmark')%>%
  summarise_each(funs(mean)) %>%
  mutate(y=Rate,x=meanTempAir_9T) %>%
  dplyr::select(-meanTempAir_9T, -Rate)

eq <- ddply(df,.(Cultivar),lm_eqn)

yx<-quantile(df$y,0.75)
xn<- quantile(df$x,0.25)

#get statistical summary here 
summary(lm(y ~ x, df))

p <- ggplot(data = df, aes(x = x, y = y)) +
  geom_smooth(method = "lm", se=FALSE, color="black", formula = y ~ x) +
  geom_point()+
  theme_bw()+
  scale_x_continuous(limits = c(0,25))+
  scale_y_continuous(limits = c(0,0.045))

p1 <- p + geom_text(data=eq,aes(x = xn, y = yx,label=V1), parse = TRUE, inherit.aes=FALSE) + 
  facet_grid(Cultivar~.,scales="free")
  
p1

```

  
  
  
  
This next script is not necessary yet. It was used to split the emergence data. 


I needed to get the equations for the interval between certain temperatures use the codes below 
5 - 17 degrees.
This is the portion of the data which is ascending.

```{r}

# # 1/DAS50E (going up)
#   g1 <- df_H %>%
#     filter(SowTreat!="S1")%>%
#   ggplot(aes(x=meanTemp_Em, y=InvDAS50E, shape=Cultivar))+
#   geom_point(colour= "red", alpha=0.2)+
#   #geom_smooth(colour="red")+
#   #stat_smooth(level = 0.9, size = 1, colour="white") + 
#   #geom_smooth(method = "lm")+
#   geom_smooth(method = "lm", se=FALSE, color="blue", formula = y ~ x) +
#             geom_point()
# 
# 
# p1 = g1 + geom_text(data=eq,aes(x = 5, y = 0.25,label=V1), parse = TRUE, inherit.aes=FALSE) + #facet_grid(group~.)
# p1 
#   theme_bw() +
#   xlab("Mean Soil Temp (oC)") + 
#   ylab(" Rate (1/ Days) to 50% emergence") +
#   facet_grid(.~Cultivar, scales="free")+
#   theme(text = element_text(size=10))+
#  scale_x_continuous(limits = c(0,17))+
#  scale_y_continuous(limits = c(0,0.25))+
#  
#    
# 
#   #grid.arrange(g1, g2)
# # 
# 
#   
#   grid.arrange(g1, g2)




```

End of this script move to Stats in script 6
