==> TrifosNineI2/10_TrifoIv2_MeanTTagainstMeanPpCalc9Trifo.Rmd <==
---
title: "Include the mean Photoperiod ""
author: "CT"
date: "16.07.2017"
output:
  pdf_document: default
  html_document: default
  word_document: default
note: This is for DAE and TT from Emergence  to 9 Trifos. Include the mean Pp. 
---

```{r loadLibraries, echo=FALSE, include=FALSE}
library(tidyr)
library (dplyr)
library (ggplot2)
library(agricolae)
library(knitr)
library (lubridate)
library (segmented)
library(hydroGOF)
library (markdown)
library (knitr)

```
 
AIM:

- Calculates the Mean Photoperiod 
- Plot TT to 50% plants  with 9 trifoliates  and plot against Mean Photoperiod

```{r   loadFile}
setwd("C:\\GitHubRepos\\SubCloverPhD\\SubClover\\TrifosNineI2")

getwd()


```
 
```{r}
#get dataframe created similar to ReAnalysis4 which is Rea6 - use here the dataframe with the emergence info : DaysTT9TFromEmergence
#df_tt_flow <- read.table("df_tt_50p_flow.txt",header=TRUE) 

df_tt_9T <- read.table("df_Pptargets_TT9Trifos.txt",header=TRUE) 

# force date format
df_tt_9T$SowingDate <- ymd(df_tt_9T$SowingDate)
str(df_tt_9T)


```

Clean table with selected columns

```{r}
#summary(df_tt_9T)
df_Select <-df_tt_9T %>%
  ##Calculate Date when 9 Trifos occurred 
  mutate(DateTo9Trifo= SowingDate+DAS9T )%>%

  dplyr::select(SowingDate:Block,DAS9TFromEmerg,TT9TFromEmerg,Pp_sow,SowingD,DateTo9Trifo)

```


#Incorporate the mean Pp as well 


## Read the daily Pp and  adjust formats 

```{r}

#create file
df_P <- read.table("df_CumTTiv.txt",header=TRUE)

#
str(df_P)
df_P$Date<-as.Date(df_P$Date)
str(df_P)

#adjust formats
df_P<-df_P%>%
 #mutate(Date=dmy(Date))%>%
  select(Date,Pp)

str(df_P)

```
 
# Now need to calculate the mean Pp

#Till here all good ! 

Need to force a date for 50 % of plants with flowers  (round) to calculate the average Pp during this phase  

Save dataframe as txt file.


```{r}


# force formats of dates
df_Select$Date <- round(df_Select$DateTo9Trifo ,0)
df_P$Date <- round(df_P $Date,0)

#then repeat same procedure done for the soing date but use the dat for 1 Trifoliate
  df_mergePp <- merge(df_Select, df_P, by="Date") 
  
  str(df_mergePp)


```




Calculate the mean photoperiod for the 9 trifos  period

```{r}


  df_Pp509TmeanTPp <-df_mergePp %>%
    mutate(meanPp509T = (Pp+Pp_sow)/2) %>%
  ##calculate the rate to Flowering 1/DAS50FFromEmerg
  mutate(Rate9T= 1/DAS9TFromEmerg)

str(df_Pp509TmeanTPp)

```


then select columns of interest and clean data frame for just 144 plots dataset  

```{r}

b <-df_Pp509TmeanTPp %>%
  select(SowingDate:Block,TT9TFromEmerg,Rate9T,meanPp509T )%>%
  group_by(SowTreat,Plot,Cultivar,Block)%>%
  dplyr::select(meanPp509T, TT9TFromEmerg,Rate9T)%>%
  summarise_all(funs(mean))
  

```





## Data Visualisation

```{r, fig.height=5, fig.width=15}


dfGraph3 <- b %>%
  na.omit() %>%
  dplyr::select(Cultivar, SowTreat, TT9TFromEmerg, meanPp509T) %>%
  group_by(Cultivar, SowTreat) %>%
  summarise_all(funs(mean,sd))


dfGraph3 %>%
mutate (SowTreat= factor(SowTreat, levels=c("S1", "S2", "S3", "S4", "S5", "S6", "S7","S8"),labels=c("June", "July", "Sept.", "Nov.", "Dec.", "Feb.", "March", "May")))%>%
   #filter(Cultivar=="Monti"| Cultivar=="Antas")%>%
  #filter (SowTreat== "S3"| SowTreat== "S4"|SowTreat== "S5"|SowTreat== "S6")%>%
  #ggplot(aes(x=Pp_sow,y=TT50F,colour=factor(Cultivar)))   +
   #ggplot(aes(x=Pp_sow,y=TT9TFromEmerg))+
ggplot(aes(x=meanPp509T_mean,y=TT9TFromEmerg_mean)) +
  #  geom_point(aes(shape= SowTreat), size=4)+
  
  geom_text(aes(label=SowTreat),hjust=0, vjust=0,size=5) +
 # geom_smooth()+
  theme_bw()+
  #remove grid lines 
   theme(
     panel.border = element_rect(colour = "black", size = 2), 
     panel.grid.major = element_blank(),
     panel.grid.minor = element_blank(),
     axis.line = element_line(colour = "black"))+
    ylab(expression(Mean~Thermal~time~to~Nine~Trifoliates~"("^"o"*"Cd)")) + 
  
  xlab(" Mean Photoperiod (hours)") +
  facet_grid(.~Cultivar, scales="free")+
  theme(text = element_text(size=18))+
 scale_x_continuous(limits = c(9,20))+
  scale_y_continuous(limits = c(0,900))
#theme(axis.title.y = element_text(face="bold", size=16))+
   

#ggsave("plot_PhotoperiodSowing.png", width=16, height=6, dpi=400)

```

End of this script
