==> TrifosNineI2/11_BTrifoIv2_CalcRegressions_Phyllochron.Rmd <==
---
title: "Trifoliates Iversen "
author: "CT"
date: "31.08.2016"
output: 
  word_document: default
  pdf_document: default
  html_document: default
---

Purpose:

-Analysis of phyllochron  
- Get regressions form Dataphyllol
 
```{r loadLibraries}
library(tidyr)
library (dplyr)
library (ggplot2)
library(agricolae)
library (lubridate)
library(mosaic)
library(lattice)
library(knitr)
```



```{r}
setwd("C:\\GitHubRepos\\SubCloverPhD\\SubClover\\TrifosNineI2")
getwd()
#create file
#df_trifo <- read.table("TrifoCountall.txt",header=TRUE)
#df_trifo <- read.table("TrifoCountall.txt",header=TRUE)
head(df_trifo)


```
Define formats and calculate the mean number of trifoliates considering the sub samples (S1-S10)

```{r loadFile}


# define factors for the ones that "look like" values
df_trifo$Plot <- as.factor(df_trifo$Plot)
df_trifo$Block <- as.factor(df_trifo$Block)
df_trifo$Cultivar <- as.factor(df_trifo$Cultivar)
df_trifo$SowTreat <- as.factor(df_trifo$SowTreat)

 df_trifo <- df_trifo %>%
  mutate(SowingDate = dmy(SowingDate), 
         ReadingDate = dmy(ReadingDate),
         Trifo = (s1+s2+s3+s4+s5+s6+s7+s8+s9+s10)/10)

str(df_trifo)
head(df_trifo)
tail(df_trifo)


df_trifo %>% filter(Plot ==1) %>% summary()



```

Graph the results : DAS and number of trifoliates 

```{r, fig.height=8, fig.width=15}
#Graph 1        
  df_trifo %>%
  mutate (SowTreat= factor(SowTreat, levels=c("S1", "S2", "S3", "S4", "S5", "S6","S7", "S8"),labels=c("June", "July", "September", "November", "December", "February","March", "May")))%>%
  filter(Cultivar=="Antas"|Cultivar=="Denmark"|Cultivar=="Leura")%>%
  group_by(Cultivar, SowTreat, DAS) %>%
  summarise_each(funs(mean)) %>%
  ggplot(aes(x=DAS, y=Trifo)) + 
  #geom_point(aes(colour=Cultivar, shape=Cultivar), size=4) +
  geom_point(aes(shape=Cultivar), size=5) +
  #scale_shape_manual(values = c(0,1,2, 3,4,5,6,7,8,9,10,11))+
  scale_shape_manual(values = c(1,2,0)) +
  #geom_line(aes(colour=Cultivar)) +
  #geom_line()+
  #facet_wrap(~SowTreat,ncol=4, scales = "free") +
  facet_wrap (~SowTreat,ncol=4)+
  ylim(0,16)+
  labs(y="Number of trifoliate leaves (n)",x="Days after sowing")+
  #theme_grey(base_size = 16) +
  theme_bw(base_size = 24)+
  
   #remove grid lines 
     theme(
     panel.border = element_rect(colour = "black", size = 2), 
     panel.grid.major = element_blank(),
     panel.grid.minor = element_blank(),
     axis.line = element_line(colour = "black"))+
  
    geom_abline(intercept = 1, slope = 0, linetype=2)+
   geom_abline(intercept = 5,  slope = 0, linetype=2)+
   geom_abline(intercept = 9,  slope = 0, linetype=2)+
theme(axis.text.x=element_text(angle = 0, hjust = 0))+
theme(text = element_text(size = 24))


#save
   
   #ggsave("plot_NumberOfTrifoliateLeavesA.png", width=16, height=8, dpi=400)




```


Graph the results : DAS and number of trifoliates 

```{r, fig.height=8, fig.width=15}
#Graph 1        
  df_trifo %>%
  mutate (SowTreat= factor(SowTreat, levels=c("S1", "S2", "S3", "S4", "S5", "S6","S7", "S8"),labels=c("June", "July", "September", "November", "December", "February","March", "May")))%>%
  filter(Cultivar=="Monti"|Cultivar=="Narrikup"|Cultivar=="Woogenellup")%>%
  group_by(Cultivar, SowTreat, DAS) %>%
  summarise_each(funs(mean)) %>%
  ggplot(aes(x=DAS, y=Trifo)) + 
  #geom_point(aes(colour=Cultivar, shape=Cultivar), size=4) +
  geom_point(aes(shape=Cultivar), size=5) +
  #scale_shape_manual(values = c(0,1,2, 3,4,5,6,7,8,9,10,11))+
  scale_shape_manual(values = c(4,7,8)) +
  #geom_line(aes(colour=Cultivar)) +
  #geom_line()+
  #facet_wrap(~SowTreat,ncol=4, scales = "free") +
  facet_wrap (~SowTreat,ncol=4)+
  ylim(0,16)+
  labs(y="Number of trifoliate leaves (n)",x="Days after sowing")+
  
  theme_bw(base_size = 24)+
  
   #remove grid lines 
     theme(
     panel.border = element_rect(colour = "black", size = 2), 
     panel.grid.major = element_blank(),
     panel.grid.minor = element_blank(),
     axis.line = element_line(colour = "black"))+
  
    geom_abline(intercept = 1, slope = 0, linetype=2)+
   geom_abline(intercept = 5,  slope = 0, linetype=2)+
   geom_abline(intercept = 9,  slope = 0, linetype=2)+
theme(axis.text.x=element_text(angle = 0, hjust = 0))+
theme(text = element_text(size = 24))

#save
   
   #ggsave("plot_NumberOfTrifoliateLeavesB.png", width=16, height=8, dpi=400)
 

```





Other graphics with same dat 
```{r, fig.height=5, fig.width=8}
#Graph         
  df_trifo %>%  
  mutate (SowTreat= factor(SowTreat, levels=c("S1", "S2", "S3", "S4", "S5", "S6","S7", "S8"),labels=c("June", "July", "September", "November", "December", "February","March", "May")))%>%
  group_by(Cultivar, SowTreat, DAS) %>%
  summarise_each(funs(mean)) %>%
  ggplot(aes(x=DAS, y=Trifo)) + 
  geom_point(aes(colour=Cultivar, shape=Cultivar)) +
  geom_line(aes(colour=Cultivar)) +
  facet_grid (SowTreat~., scales = "free")+
  labs(y="Number of trifoliates",x="Days after sowing")+
    #geom_abline(intercept = 6, slope = 0)
   geom_abline(intercept = 4:6, slope = 0)



```


get the standard  error values for the final trifoliate numbers 


```{r}
str(df_trifo)

dfMaxTrifo <- df_trifo %>% 

  group_by( Cultivar, SowTreat ) %>%
  dplyr::select(Trifo)%>%
  summarise_all(funs(max,sd))%>%
  mutate(SE=sd/1.73)
 

  



```


Then include the TT calculated !!! TT cum 




Read Thermal Sum data. Need to include  values before 30 06 : 

Country	LocationYear	Date	DOY	Tx	Tm	MeanTt	Pp
NZ	Iversen2_2015	24/06/2015	175	7.30	-1.80	2.75	10.03173957
NZ	Iversen2_2015	25/06/2015	176	6.90	-3.90	1.50	10.033705
NZ	Iversen2_2015	26/06/2015	177	12.80	-0.70	6.05	10.03668323
NZ	Iversen2_2015	27/06/2015	178	17.40	-1.30	8.05	10.04067231
NZ	Iversen2_2015	28/06/2015	179	10.30	-1.70	4.30	10.04566961
NZ	Iversen2_2015	29/06/2015	180	17.30	3.70	10.50	10.05167184
NZ	Iversen2_2015	30/06/2015	181	18.50	1.30	9.90	10.05867507
"

```{r ReadThermalSumData}
# read raw data

#Temperatures from 24 June to 29 June included from Broadfield Niwa in excel Datalogger folder  Thermal Sum is created  in the the Thermal time project_TTCalc (R script_file)

#The df_CumTTiv.txt comes from the TTCalc script in the main Gitsubclover folder! Need to produce the file and add to the FlowIv2 folder for calculations. 

#the file df_CumTTiv2.txt contains the thermal sum hourly from Iversen. 
 
df_TTSum <- read.table("df_CumTTiv2.txt",header=TRUE)

# remove extra columns
df_TTSum <- df_TTSum %>%
dplyr::select(Date,SumTTday) %>%
  mutate(Date = ymd(Date))
 
#check
head(df_TTSum)

```

Check data 
```{r}
summary(df_TTSum)
```



Merge dfs TTSum at Measurement and Sowing Date into Flowering df
```{r}

str(df_trifo)
df_trifo<-df_trifo%>%
  mutate(Date=ymd(ReadingDate))
str(df_trifo)

# Add Tt at measurement date
df_trifo2 <- merge(df_trifo,df_TTSum, by="Date")

summary(df_trifo2)

```


Change column names and select only columns of interest (tidy up)

```{r}


# change name SumTTday to TT at measurement used mutate 1 x cause rename did not work 
df_trifo2 <- df_trifo2 %>%
mutate(TT_meas = 1*(SumTTday))
#check the data 
str(df_trifo2)

#select only columns of interest 
df_trifo3<- df_trifo2 %>%
dplyr::select (-s1, -s2, -s3, -s4, -s5,-s6, -s7, -s8, -s9, -s10) 

#check data 
str(df_trifo3)
```

Need to modify the column name to be able to merge 
```{r}
# Add Tt at sowing date
df_TTSum_mod <- df_TTSum %>%
  mutate (SowingDate = Date) %>%
  dplyr::select (-Date)

 str(df_TTSum_mod)
 str(df_trifo3)

```


Merge data frames 
```{r}



df_trifo4 <- merge(df_trifo3,df_TTSum_mod, by = "SowingDate")

str(df_trifo4)

# change name TT to TT at measurement
df_trifo4 <- df_trifo4 %>%
mutate(TT_sow = 1*(SumTTday.y))
#check
str(df_trifo4)

```


Graphic

```{r MergeTTSumDate}

#aqi
# TT from Bud to Bur 3 

df_trifo4 %>%
  #filter(SowTreat != "S6" & SowTreat != "S7")%>%
  group_by(DAS,Cultivar,SowTreat) %>%
  summarise_each(funs(mean)) %>%
  #select(Variable,Percent, Cultivar,SowTreat) %>%
  filter(Cultivar == "Narrikup" | Cultivar =="Denmark") %>%
  ggplot(aes(x=DAS, y=Trifo)) + 
  #geom_bar(shape=1, stat="identity")
  geom_line(aes(colour=Cultivar)) +
  facet_grid(SowTreat~., scales = "free") +
  geom_abline(intercept = 50, slope = 0)+
  
theme(axis.text.x = element_text(angle = 90, hjust = 1))


```

calculate TT after sowing 

```{r Calc_TT_afterSow}

df_trifo4$TTAS <- df_trifo4$TT_meas-df_trifo4$TT_sow
head(df_trifo4)


summary(df_trifo4)



```


Include Photoperiod at sowing date in main df 

```{r mergePhotoperiod}

# #create file
df_Photo <- read.table("Photoperiod.txt",header=TRUE)

# convert to date format
df_Photo <- df_Photo %>%
mutate(Date=dmy(Date))

str(df_Photo)

# rename date column for merge
df_Photo <- df_Photo %>%
mutate(SowingDate = Date)
head(df_Photo)

#Add Pp to df_flower (merge by SowinD)
df_trifo4 <- merge(df_trifo4,df_Photo,by="SowingDate")
head(df_trifo4)
summary(df_trifo4)


 # change name PP to PP at sowing date
df_trifo4 <- df_trifo4 %>%
mutate(Pp_sow = Pp)
str(df_trifo4)



```



```{r}

str(df_trifo4)

# save final DF"
write.table(df_trifo4, "df_TrifoLeavesTT.txt")

```

Graphics 

```{r}
# graph
df_trifo4 %>%
  # mutate(Variable = 
  # factor(Variable,levels = c("Bud","Early","Open",
  #         "Petal","Bur1","Bur2","Bur3","Bur4" ))) %>%
  filter(Cultivar=="Narrikup")%>%
  group_by(TTAS,Cultivar,SowTreat,DAS) %>%
  summarise_each(funs(mean)) %>%
  #select(Variable,Percent, Cultivar,TTAS,SowTreat,SowingD,DAS) %>%
  #filter(Variable == "Bud"| Variable == "Open" ) %>%
  #filter(Variable == "Bud"| Variable == "Bur3" ) %>%
  #ggplot(aes(x=TTAS, y=Percent)) + 
  ggplot(aes(x=TTAS, y=Trifo)) +
  geom_line(aes(colour=SowTreat)) +
  facet_grid(SowTreat~Cultivar, scales = "free") +
  geom_abline(intercept = 50, slope = 0)+
   theme_bw()+
theme(axis.text.x = element_text(angle = 60, hjust = 1))+
labs(y="Percent of Plants with Floral bud ")+
xlab(expression(Thermal~time~~"("^"o"*"Cd)"))
 
```

Regraph

Graph the results : DAS and number of trifoliates 

```{r, fig.height=8, fig.width=15}
#Graph 1        
  df_trifo4 %>%
  mutate (SowTreat= factor(SowTreat, levels=c("S1", "S2", "S3", "S4", "S5", "S6","S7", "S8"),labels=c("June", "July", "September", "November", "December", "February","March", "May")))%>%
  filter(Cultivar=="Monti"|Cultivar=="Narrikup"|Cultivar=="Woogenellup")%>%
  group_by(Cultivar, SowTreat, DAS,TTAS) %>%
  summarise_each(funs(mean)) %>%
  ggplot(aes(x=TTAS, y=Trifo)) + 
  #geom_point(aes(colour=Cultivar, shape=Cultivar), size=4) +
  geom_point(aes(shape=Cultivar), size=5) +
  #scale_shape_manual(values = c(0,1,2, 3,4,5,6,7,8,9,10,11))+
  scale_shape_manual(values = c(4,7,8)) +
  #geom_line(aes(colour=Cultivar)) +
  #geom_line()+
  #facet_wrap(~SowTreat,ncol=4, scales = "free") +
  facet_wrap (~SowTreat,ncol=4)+
  ylim(0,16)+
  labs(y="Number of trifoliate leaves (n)")+
  xlab(expression(Thermal~time~Sowing~"("^"o"*"Cd)"))+
  theme_bw(base_size = 24)+
  
   #remove grid lines 
     theme(
     panel.border = element_rect(colour = "black", size = 2), 
     panel.grid.major = element_blank(),
     panel.grid.minor = element_blank(),
     axis.line = element_line(colour = "black"))+
  
    geom_abline(intercept = 1, slope = 0, linetype=2)+
   geom_abline(intercept = 5,  slope = 0, linetype=2)+
   geom_abline(intercept = 9,  slope = 0, linetype=2)+
theme(axis.text.x=element_text(angle = 0, hjust = 0))+
theme(text = element_text(size = 24))

#save
   
   #ggsave("plot_NumberOfTrifoliateLeavesB.png", width=16, height=8, dpi=400)
 

```





bring dataframe with TTEmergence to calculate number leaves from emergence



```{r}

df_TTEmerg <- read.table("DaysTT9TFromEmergence_v2.txt",header=TRUE)

str(df_TTEmerg) 


df_TTEmerg <- df_TTEmerg%>%
  dplyr::select(Plot,SowTreat,SowingDate,Cultivar,Block,TT50E,DAS50E)

##merge 
df_TTEmerg2<-merge(df_TTEmerg,df_trifo4, by="Plot")

##subtract TT
df_TTEmerg2$TTAEmer <- df_TTEmerg2$TTAS-df_TTEmerg2$TT50E
head(df_TTEmerg2)


summary(df_TTEmerg2)
str(df_TTEmerg2)


```

tidy df 

```{r}

df_TTEmerg2<-df_TTEmerg2 %>%
  mutate( SowTreat=SowTreat.x,
          Cultivar=Cultivar.x, 
          Block=Block.x,
          SowingDate=SowingDate.x)%>%
  select(-SowTreat.y,-Cultivar.y,- Block.y , -SumTTday.y)

  

```

Select only data needed for regressions ans stats 
```{r}

##this will be used to calculate the slope (phyllochron)

DataPhyllo<-df_TTEmerg2%>%
  dplyr::select(Plot,Block,SowTreat,Cultivar,SowingDate,Pp_sow, TTAEmer,Trifo)%>%
  subset(TTAEmer>0)%>%
  subset(Trifo>0.65 & Trifo<6)

write.table(DataPhyllo, "DataPhyllo.txt")
```



```{r, Warning=FALSE}

df_1 <-df_TTEmerg2%>%
  group_by(Cultivar, SowTreat,DAS)%>%
  dplyr::select(Trifo,TTAEmer) %>%
  summarise_all(funs(mean,sd))%>%
  subset(TTAEmer_mean>0)


#str(df_1)

df_1select<-df_1%>%
  select (Cultivar,SowTreat,TTAEmer_mean,Trifo_mean,Trifo_sd)%>%
  mutate(TT=TTAEmer_mean,
         mean=Trifo_mean,
         sd=Trifo_sd)%>%
  select(-TTAEmer_mean,-Trifo_mean,-Trifo_sd)
  
  str(df_1select)
  
 write.table(df_1select, "df_1select.txt")
    
  
```

make graph with all cultivars 



```{r, fig.height=8, fig.width=15}

##Find the maximum SE 
##Create datapoint for error bar for the final number leaves
dfError<- df_1select %>%
  subset(mean>8.5)%>%
  #dplyr::select(SowTreat,sd)%>%
  group_by(SowTreat)%>%
  #dplyr::select(sd)%>%
  summarise_all(funs(max(sd/1.73)))%>%
  mutate(SE=sd, TT=650, mean=15)%>% 
  mutate (SowTreat= factor(SowTreat, levels=c("S1", "S2", "S3", "S4", "S5", "S6","S7", "S8"),labels=c("June", "July", "September", "November", "December", "February","March", "May")))%>%
  ungroup()


###########################################
## create a dataframewitth values of trifo >0.8 and < 6
TrifoSubset<-df_1select%>%
  subset(mean>=0.8 & mean <= 7)%>%
  mutate (SowTreat= factor(SowTreat, levels=c("S1", "S2", "S3", "S4", "S5", "S6","S7", "S8"),labels=c("June", "July", "September", "November", "December", "February","March", "May")))


#Graph 1  
#str(df_1select)
  df_1select %>%
    ungroup()%>%
  mutate (SowTreat= factor(SowTreat, levels=c("S1", "S2", "S3", "S4", "S5", "S6","S7", "S8"),labels=c("June", "July", "September", "November", "December", "February","March", "May")))%>%
    subset(mean>0.50)%>%
  #filter(Cultivar=="Antas"|Cultivar=="Denmark"|Cultivar=="Leura")%>%
  #group_by(Cultivar, SowTreat, DAS) %>%
  #summarise_each(funs(mean)) %>%
  ggplot(aes(x=TT, y=mean)) + 
  geom_smooth(data=TrifoSubset,method='lm',formula=y~x, colour="grey",alpha=0.1)+
    
  geom_point(aes(shape=Cultivar), size=5) +
  #scale_shape_manual(values = c(0,1,2, 3,4,5,6,7,8,9,10,11))+
  scale_shape_manual(values = c(0,1,2,4,7,8)) +
    geom_errorbar(data=dfError, aes(ymin=mean-SE,ymax=mean+SE), width=50)+
  #geom_line(aes(colour=Cultivar)) +
  #geom_line()+
  #facet_wrap(~SowTreat,ncol=4, scales = "free") +
  facet_wrap (~SowTreat,ncol=4)+
  ylim(0,18)+
  xlim(0,700)+
  labs(y="Number of trifoliate leaves (n)")+
  #xlab(expression(Thermal~time~Emergence~"("^"o"*"Cd)"))+
  xlab(expression(Thermal~time~"("^"o"*"Cd)"))+
  theme_bw(base_size = 24)+
  
   #remove grid lines 
     theme(
     panel.border = element_rect(colour = "black", size = 2), 
     panel.grid.major = element_blank(),
     panel.grid.minor = element_blank(),
     axis.line = element_line(colour = "black"))+
  
    geom_abline(intercept = 1, slope = 0, linetype=2)+
   geom_abline(intercept = 5,  slope = 0, linetype=2)+
   #geom_abline(intercept = 9,  slope = 0, linetype=2)+
theme(axis.text.x=element_text(angle = 0, hjust = 0.5))+
theme(text = element_text(size = 24))+
    ##thicks in axis
  theme(axis.ticks.length=unit(.3, "cm"))+
    #make Cultivars  white  
   theme(strip.background = element_rect(colour = "black", fill = "white",size=2))+
  
  #label sizes 
theme(text = element_text(size = 26))

#save
   
   ggsave("plot_11_AllCVS_NumberOfTrifoliateLeavesThermalT.tiff", width=24, height=12, dpi=400)
 

```



make graph A, D,L



```{r, fig.height=8, fig.width=15}
#Graph 1  
str(df_1select)
  df_1select %>%
    ungroup()%>%
  mutate (SowTreat= factor(SowTreat, levels=c("S1", "S2", "S3", "S4", "S5", "S6","S7", "S8"),labels=c("June", "July", "September", "November", "December", "February","March", "May")))%>%
  filter(Cultivar=="Antas"|Cultivar=="Denmark"|Cultivar=="Leura")%>%
  #group_by(Cultivar, SowTreat, DAS) %>%
  #summarise_each(funs(mean)) %>%
  ggplot(aes(x=TT, y=mean)) + 
  #geom_point(aes(colour=Cultivar, shape=Cultivar), size=4) +
  geom_point(aes(shape=Cultivar), size=5) +
  #scale_shape_manual(values = c(0,1,2, 3,4,5,6,7,8,9,10,11))+
  scale_shape_manual(values = c(0,1,2)) +
  #geom_line(aes(colour=Cultivar)) +
  #geom_line()+
  #facet_wrap(~SowTreat,ncol=4, scales = "free") +
  facet_wrap (~SowTreat,ncol=4)+
  ylim(0,16)+
  labs(y="Number of trifoliate leaves (n)")+
  #xlab(expression(Thermal~time~Emergence~"("^"o"*"Cd)"))+
  xlab(expression(Thermal~time~"("^"o"*"Cd)"))+
  theme_bw(base_size = 24)+
  
   #remove grid lines 
     theme(
     panel.border = element_rect(colour = "black", size = 2), 
     panel.grid.major = element_blank(),
     panel.grid.minor = element_blank(),
     axis.line = element_line(colour = "black"))+
  
    geom_abline(intercept = 1, slope = 0, linetype=2)+
   geom_abline(intercept = 5,  slope = 0, linetype=2)+
   geom_abline(intercept = 9,  slope = 0, linetype=2)+
theme(axis.text.x=element_text(angle = 0, hjust = 0.5))+
theme(text = element_text(size = 24))+
    ##thicks in axis
  theme(axis.ticks.length=unit(.3, "cm"))+
    #make Cultivars  white  
   theme(strip.background = element_rect(colour = "black", fill = "white",size=2))+
  
  #label sizes 
theme(text = element_text(size = 26))

#save
   
   ggsave("plot_11_NumberOfTrifoliateLeavesThermalT_ADL.tiff", width=16, height=8, dpi=400)
 

```




```{r, fig.height=8, fig.width=15}
#Graph 1  
#str(df_1select)
  df_1select %>%
    ungroup()%>%
  mutate (SowTreat= factor(SowTreat, levels=c("S1", "S2", "S3", "S4", "S5", "S6","S7", "S8"),labels=c("June", "July", "September", "November", "December", "February","March", "May")))%>%
  filter(Cultivar=="Monti"|Cultivar=="Narrikup"|Cultivar=="Woogenellup")%>%
  #group_by(Cultivar, SowTreat, DAS) %>%
  #summarise_each(funs(mean)) %>%
  ggplot(aes(x=TT, y=mean)) + 
  #geom_point(aes(colour=Cultivar, shape=Cultivar), size=4) +
  geom_point(aes(shape=Cultivar), size=5) +
  #scale_shape_manual(values = c(0,1,2, 3,4,5,6,7,8,9,10,11))+
  scale_shape_manual(values = c(4,7,8)) +
  #geom_line(aes(colour=Cultivar)) +
  #geom_line()+
  #facet_wrap(~SowTreat,ncol=4, scales = "free") +
  facet_wrap (~SowTreat,ncol=4)+
  ylim(0,16)+
  labs(y="Number of trifoliate leaves (n)")+
  xlab(expression(Thermal~time~"("^"o"*"Cd)"))+
     #xlab(expression(Thermal~time~Emergence~"("^"o"*"Cd)"))+
  theme_bw(base_size = 24)+
  
   #remove grid lines 
     theme(
     panel.border = element_rect(colour = "black", size = 2), 
     panel.grid.major = element_blank(),
     panel.grid.minor = element_blank(),
     axis.line = element_line(colour = "black"))+
  
    geom_abline(intercept = 1, slope = 0, linetype=2)+
   geom_abline(intercept = 5,  slope = 0, linetype=2)+
   geom_abline(intercept = 9,  slope = 0, linetype=2)+
theme(axis.text.x=element_text(angle = 0, hjust = 0.5))+
theme(text = element_text(size = 24))+
    ##thicks in axis
  theme(axis.ticks.length=unit(.3, "cm"))+
    #make Cultivars  white  
   theme(strip.background = element_rect(colour = "black", fill = "white",size=2))+
  
  #label sizes 
theme(text = element_text(size = 26))

#save
   
   ggsave("plot_11_NumberOfTrifoliateLeavesThermalT_MNW.tiff", width=16, height=8, dpi=400)
 

```

make the regressions 



This loop checks the lm for all cultivars 

```{r}


#create a vector that holds names of all possible cultivars
cv.unique <- unique(DataPhyllo$SowTreat)

summary(cv.unique )


for (i in 1:length(cv.unique)) {
  
 #print(i)
 print(paste0(cv.unique[i],"--------------------------------"))
  
  #then make it subset 
  
  df_subset <- DataPhyllo %>%
    subset(SowTreat==cv.unique[i])
  
  #add the function or calculation required 
  
    #summary(lm(HardAccum ~ RainIrrig, df_subset))  
    #summary(lm(HardAccum ~ RH, df_subset))
    summary(lm(Trifo ~ TTAEmer, df_subset))
  #print
    
print(summary(lm(Trifo ~ TTAEmer, df_subset)))
#print(summary(lm(HardAccum ~ RH, df_subset)))
   # print(summary(lm(HardAccum ~ Temp, df_subset)))

  } 


```
